fix(keys): preserve verification across cosmetic refreshes
