+++ tests/entrypoints/options/pages/KeyManagement/TokenHeader.saveToApiProfiles.test.tsx
@@ -1,5 +1,5 @@
 import userEvent from "@testing-library/user-event"
-import type { ComponentProps, ReactNode } from "react"
+import type { ReactNode } from "react"
 import toast from "react-hot-toast"
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
@@ -8,10 +8,6 @@ import {
   AIHUBMIX_WEB_ORIGIN,
   SITE_TYPES,
 } from "~/constants/siteType"
-import { TokenHeader as KeyResourceTokenHeader } from "~/features/KeyManagement/components/TokenListItem/TokenHeader"
-import type { KeyResourceActionPolicy } from "~/features/KeyManagement/presentation/keyResourceCard"
-import { buildLegacyKeyResourceCardPresentation } from "~/features/KeyManagement/presentation/legacyKeyResourceCard"
-import { buildDisplayAccountTokenRuntimeKey } from "~/services/accounts/accountRuntimeKeys"
 import {
   MANAGED_SITE_CHANNEL_KEY_MATCH_REASONS,
   MANAGED_SITE_CHANNEL_MODELS_MATCH_REASONS,
@@ -25,6 +21,10 @@ import { API_TYPES } from "~/services/verification/aiApiVerification"
 import { AuthTypeEnum, SiteHealthStatus, type DisplaySiteData } from "~/types"
 import { buildCompleteTodayStatsAvailability } from "~~/tests/test-utils/accountTodayStats"
 import { testI18n } from "~~/tests/test-utils/i18n"
+import {
+  RECOVERABLE_ACTION_POLICY,
+  TokenHeaderHarness as TokenHeader,
+} from "~~/tests/test-utils/keyManagement/TokenHeaderHarness"
 import { render, screen, waitFor } from "~~/tests/test-utils/render"
 
 const mockCreateProfile = vi.fn()
@@ -34,16 +34,6 @@ const mockOpenManagedSiteChannelsPage = vi.fn()
 const mockOpenSettingsTab = vi.fn()
 const mockOpenWithAccount = vi.fn()
 const mockLoggerError = vi.fn()
-
-const RECOVERABLE_ACTION_POLICY: KeyResourceActionPolicy = {
-  copySecret: true,
-  revealSecret: true,
-  verifySecret: true,
-  exportSecret: true,
-  edit: true,
-  delete: true,
-  batchSelect: true,
-}
 
 vi.mock(
   "~/services/apiCredentialProfiles/apiCredentialProfilesStorage",
@@ -132,28 +122,6 @@ function createAccountStub(): DisplaySiteData {
     authType: AuthTypeEnum.AccessToken,
     checkIn: { enableDetection: false },
   }
-}
-
-type TokenHeaderTestProps = Omit<
-  ComponentProps<typeof KeyResourceTokenHeader>,
-  "headerProps" | "actionPolicy"
-> & {
-  actionPolicy?: KeyResourceActionPolicy
-}
-
-function TokenHeader({ actionPolicy, ...props }: TokenHeaderTestProps) {
-  const presentation = buildLegacyKeyResourceCardPresentation(
-    buildDisplayAccountTokenRuntimeKey(props.account, props.token),
-    testI18n.t,
-  )
-
-  return (
-    <KeyResourceTokenHeader
-      {...props}
-      headerProps={{ presentation, detailsTrigger: null }}
-      actionPolicy={actionPolicy ?? presentation.actions}
-    />
-  )
 }
 
 /**
